requirements-cpu.txt
# SDRA-Rural CPU-Only Requirements
# ==================================
# For 8GB RAM, CPU-only operation
# Saves ~2GB disk space vs CUDA version

# Core
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
torchvision==0.16.0+cpu

# PDF Processing
pymupdf>=1.23.0
Pillow>=9.0.0

# OCR
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# Optional: Surya OCR (high quality)
# surya-ocr>=0.4.0

# Optional: Florence-2 VLM
transformers>=4.35.0

# Text Processing
rapidfuzz>=3.0.0

# Image Processing
opencv-python-headless>=4.8.0
numpy>=1.24.0

# Memory Monitoring
psutil>=5.9.0

# Database
duckdb>=0.9.0

# GUI
ttkbootstrap>=1.10.0

# Drag and Drop (optional)
# tkinterdnd2>=0.3.0
